refactor: boxes/messages: Extract MessageBox into new messages file.
Imports become simplified in these files due to only including relevant imports.

Imports and tests updated in additional files due to code move.

Docstrings for boxes.py and messages.py improved.
developer-file-overview.md regenerated.
